ref: move utils to "main.rs"
